iterations/v2/.stryker-tmp/sandbox-kEEo5n/python-services/dspy-integration/requirements.txt
# DSPy Integration Service Requirements
# Python 3.10+ required

# Core DSPy framework
dspy-ai>=2.4.0
requests>=2.31.0

# API server for TypeScript integration
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# LLM providers
openai>=1.3.0
anthropic>=0.7.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Type checking
mypy>=1.7.0
types-redis

# Logging and monitoring
structlog>=23.2.0

